// src/types.ts
export interface UserProps {
  id?: number;
  email: string;
  address: string;
  city: string;
  state: string;
  zip: string;
  billing_name: string;
  billing_address: string;
  billing_city: string;
  billing_state: string;
  billing_zip: string;
  work_phone: string;
  home_phone: string;
  mobile_phone: string;
}